Move `NodeGroupFilter` to cmdutils
