+++ pkg/ctl/create/nodegroup_utils_test.go
@@ -1,4 +1,4 @@
-package create
+package create_test
 
 import (
 	"bytes"
@@ -7,10 +7,13 @@ import (
 	. "github.com/onsi/gomega"
 
 	api "github.com/weaveworks/eksctl/pkg/apis/eksctl.io/v1alpha4"
+	"github.com/weaveworks/eksctl/pkg/ctl/cmdutils"
 	"github.com/weaveworks/eksctl/pkg/printers"
+
+	. "github.com/weaveworks/eksctl/pkg/ctl/create"
 )
 
-var _ = Describe("create utils", func() {
+var _ = Describe("create nodegroup utils", func() {
 
 	newClusterConfig := func() *api.ClusterConfig {
 		cfg := api.NewClusterConfig()
@@ -20,6 +23,7 @@ var _ = Describe("create utils", func() {
 
 		return cfg
 	}
+
 	addGroupA := func(cfg *api.ClusterConfig) {
 		var ng *api.NodeGroup
 
@@ -265,75 +269,20 @@ var _ = Describe("create utils", func() {
 	  }
 	`
 
-	Context("CheckEachNodeGroup", func() {
-
-		It("should iterate over unique nodegroups", func() {
-			cfg := newClusterConfig()
-			addGroupA(cfg)
-			f := NewNodeGroupFilter()
-
-			names := []string{}
-			checkEachNodeGroup(f, cfg.NodeGroups, func(i int, nodeGroup *api.NodeGroup) error {
-				Expect(nodeGroup).To(Equal(cfg.NodeGroups[i]))
-				names = append(names, nodeGroup.Name)
-				return nil
-			})
-			Expect(names).To(Equal([]string{"test-ng1a", "test-ng2a", "test-ng3a"}))
-
-			names = []string{}
-			cfg.NodeGroups[0].Name = "ng-x0"
-			cfg.NodeGroups[1].Name = "ng-x1"
-			cfg.NodeGroups[2].Name = "ng-x2"
-			checkEachNodeGroup(f, cfg.NodeGroups, func(i int, nodeGroup *api.NodeGroup) error {
-				Expect(nodeGroup).To(Equal(cfg.NodeGroups[i]))
-				names = append(names, nodeGroup.Name)
-				return nil
-			})
-			Expect(names).To(Equal([]string{"ng-x0", "ng-x1", "ng-x2"}))
-		})
-
-		It("should iterate over unique nodegroups and filter some out", func() {
-			cfg := newClusterConfig()
-			addGroupA(cfg)
-			addGroupB(cfg)
-			f := NewNodeGroupFilter()
-
-			names := []string{}
-			checkEachNodeGroup(f, cfg.NodeGroups, func(i int, nodeGroup *api.NodeGroup) error {
-				Expect(nodeGroup).To(Equal(cfg.NodeGroups[i]))
-				names = append(names, nodeGroup.Name)
-				return nil
-			})
-			Expect(names).To(Equal([]string{"test-ng1a", "test-ng2a", "test-ng3a", "test-ng1b", "test-ng2b", "test-ng3b"}))
-
-			names = []string{}
-
-			err := f.ApplyOnlyFilter([]string{"t?xyz?", "ab*z123?"}, cfg)
-			Expect(err).To(HaveOccurred())
-			Expect(err.Error()).To(Equal(`no nodegroups match filter specification: "t?xyz?,ab*z123?"`))
-
-			err = f.ApplyOnlyFilter([]string{"test-ng1?", "te*-ng3?"}, cfg)
-			Expect(err).ToNot(HaveOccurred())
-			checkEachNodeGroup(f, cfg.NodeGroups, func(i int, nodeGroup *api.NodeGroup) error {
-				Expect(nodeGroup).To(Equal(cfg.NodeGroups[i]))
-				names = append(names, nodeGroup.Name)
-				return nil
-			})
-			Expect(names).To(Equal([]string{"test-ng1a", "test-ng3a", "test-ng1b", "test-ng3b"}))
-		})
+	Context("apply nodegroup defaults", func() {
 
 		It("should iterate over unique nodegroups and apply defaults with NewNodeGroupChecker", func() {
 			cfg := newClusterConfig()
 			addGroupA(cfg)
 			addGroupB(cfg)
-			f := NewNodeGroupFilter()
+			f := cmdutils.NewNodeGroupFilter()
 
 			printer := printers.NewJSONPrinter()
 
 			names := []string{}
-			checkEachNodeGroup(f, cfg.NodeGroups, NewNodeGroupChecker)
+			f.CheckEachNodeGroup(cfg.NodeGroups, NewNodeGroupChecker)
 
-			checkEachNodeGroup(f, cfg.NodeGroups, func(i int, nodeGroup *api.NodeGroup) error {
+			f.CheckEachNodeGroup(cfg.NodeGroups, func(i int, nodeGroup *api.NodeGroup) error {
 				Expect(nodeGroup).To(Equal(cfg.NodeGroups[i]))
 				names = append(names, nodeGroup.Name)
 				return nil
